Do not export library prefix names used in a library
The spec is unclear whether library prefix names are part of
the export namespace of a library or whether they are private
to that library. The latter makes more sense and is what dart2js
implements, so here's a VM change that implements the same
behavior. Spec will be clarified.

Fixes issue 12713.

R=iposva@google.com

Review URL: https://codereview.chromium.org//23613006

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@27079 260f80e4-7a28-3924-810f-c04153c831b5