fix shadowdom polyfill
I'm not sure when it broke, but it was not rendering TodoMVC at startup (actually -- it was rendering, then clearing). In any case, the bug goes away with an updated ShadowDOM. We're now completely unforked

I also fixed minify mode. It no longer depends on the name 'GeneratedWrapper', instead it detects the precise problematic case and deals with it.

R=sigmund@google.com

Review URL: https://codereview.chromium.org//24129004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@27458 260f80e4-7a28-3924-810f-c04153c831b5