Bug fix in load elimination,  constant propagation for LoadClassId.
Handle LoadClassId in constant propagation. This helps with eliminating
unreachable code after polymorphic inlining. It exposed a bug which is
also fixed in this CL.

Redefinitions of values have to be considered for load/store forwarding:
Loads/stores to an object defined by a RedefinitionInstr (inserted by
polymorphic inlining) must be treated properly as loads/stores to the
original object.

R=kmillikin@google.com

Review URL: https://codereview.chromium.org//25640004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@28159 260f80e4-7a28-3924-810f-c04153c831b5