Add a much simplified set implementation designed to waste little memory for small sets.
Use the new set (called setlet) in the type inferrer. I'd be happy to get rid of this
specialized implementation if we ever get a core library implementation that is equally
small and fast, but for now this CL reduces the working set with 15% when compiling
dart2js.

R=ahe@google.com, lrn@google.com, lry@google.com
BUG=

Review URL: https://codereview.chromium.org//27510002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@28784 260f80e4-7a28-3924-810f-c04153c831b5