Fixes a bug in runtime call stub on arm and mips.
GC was failing to visit the last argument to
runtime calls because the stack frame iterator
expects the same layout in the runtime call
stub as with other frames. Using the stub entry
macro ensures that the last argument is in the
expected place.

R=regis@google.com

Review URL: https://codereview.chromium.org//27588004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@28803 260f80e4-7a28-3924-810f-c04153c831b5