Capture and report stdout/stderr of the browser while running a test
Currently we have no way of getting stdout/stderr information when using the
browser controller.

This CL will capture stdout/stderr of the browser and tries to align it with the
currently executed test (which might be slightly off due to the fact that
Process.{stderr,stdout} is reported asynchronously).

R=ricow@google.com

Review URL: https://codereview.chromium.org//28533003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@28858 260f80e4-7a28-3924-810f-c04153c831b5