Omnibox GTK focus fix back-peddle.
In an ideal world the omnibox popup should never get focus.  However in the world that is GTK, a shell open insists on stealing focus from the text area and if we do not take it, the focus control will go to null.  In that case, nothing has keyboard focus and all key events fall on the floor.

Still investigating better approaches but this back-step at least restores basic sanity.

R=scheglov@google.com

Review URL: https://codereview.chromium.org//31263005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@29124 260f80e4-7a28-3924-810f-c04153c831b5