Re-mark directory_watcher/mac_os_test as passing.
As before, this test is expected to fail. It's only failing on the
bots so we need to get the debug information from the bots to figure
out how to fix it.

This also adds some additional debug information to help clarify some
confusing output from the most recent failing log.

R=rnystrom@google.com
BUG=15024

Review URL: https://codereview.chromium.org//64383004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@30315 260f80e4-7a28-3924-810f-c04153c831b5