Remove dead code from IfThenElse and StrictCompare in the compiler.
The code for these cases is either completely unused or can only be hit with
canonicalization and constant propagation turned off. Since this is not tested,
and does not bring any performance benefit for the default configuration it is
best to just remove it.

R=srdjan@google.com

Review URL: https://codereview.chromium.org//62203018

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@30393 260f80e4-7a28-3924-810f-c04153c831b5