VM: Generalize using propagated cids for to >1 arguments.
When ic-data are missing (optimized code that was not executed yet),
the compiler tries to fill in information calculated with type propagation.
The approach was only used for calls with 1 tested argument. This
enables it for all calls where we want to use propagated type
information if ic-data are missing.

R=srdjan@google.com

Review URL: https://codereview.chromium.org//110703004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@31020 260f80e4-7a28-3924-810f-c04153c831b5