GenerateNot was not respecting "generateAtUseSite".
This fixes a severe issue when the first condition of a logical-or was used
by more than one expression.

Example:
var cond = foo();
if (cond || bar()) ...
if (cond || gee()) ...

An intermediate step transformed this into:
var t1 = !foo();  // <== note the "!".
if (t1) if (bar()) -> jump to correct block
if (t1) if (gee()) -> jump to correct block.

The variable 't1' was marked as "not-generate-at-use-site", since it
was shared between the two 'ifs'.

Later we realize that we want to emit a logical or.

However, in the back-end we had an optimization that basically said:
if we produce a logical or, we have to invert the first argument.
if that argument is a Not, we can ignore the Not-node and use its expression
directly.

This jumped over the "not-generate-at-use-site" of the Not node, and
looked directly at "cond", which was allowed to be generated at
use-site (at the initialization of "t1").

BUG= http://dartbug.com/16996
R=sra@google.com

Review URL: https://codereview.chromium.org//175403002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@32951 260f80e4-7a28-3924-810f-c04153c831b5