Remove content cache from FileBasedSource and SourceFactory.
Sorry, again, about the size. I was right, though, that I didn't want to combine this clean-up with the previous work. (The bad news is that there's a lot more clean-up that could be done in the tests. Not sure when I'll get to it.)

R=danrubel@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org//176613002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@32970 260f80e4-7a28-3924-810f-c04153c831b5