Add a property to the initial global objects.
We want the initial objects to be in slow mode, before we "finalize" them
and transform to fast objects. We use "delete" for that, but since we
didn't have the property in the object "delete" wasn't doing anything and
kept the objects in fast mode (where we pay for the map-transitions).

R=ahe@google.com

Review URL: https://codereview.chromium.org//180443002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@33055 260f80e4-7a28-3924-810f-c04153c831b5