Change handling of dart-ext: URIs for native extensions.
With this change, a dart-ext: URI is now an opaque URI consisting of
the scheme dart-ext: followed by an actual URI, either relative, or
with a scheme like file:, http:, or package:.  Existing users are mostly
or all using relative URIs, so this change should not affect them.

This is a step toward allowing native extensions in packages to be served
over http, and also cleans up the code, and would help a future change
that replaces "dart-ext:" with metadata or an annotation.

BUG=
R=iposva@google.com

Review URL: https://codereview.chromium.org//180783008

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@33216 260f80e4-7a28-3924-810f-c04153c831b5