Nit change in the Dart command line analyzer and analysis server.
Logger.logError takes a String, not Error types.  This will enable the exception to be printed, instead of the message '<SomeError> is not a String'.

Also, in the command line version of the dart analyzer, I put in a commented out stdout/stderr logger for developers.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//196373003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@33587 260f80e4-7a28-3924-810f-c04153c831b5