Fix for queue closed crash
It is possible for JavaScriptBackend.onQueueEmpty to be called twice, first time with isTreeShakingDisabled false (hence calling Enqueuer.enqueueReflectiveStaticFields), the second time with isTreeShakingDisabled true (hence calling Enqueuer.enqueueEverything).  The crash is caused by using the same flag to avoid repeating work for enqueueReflectiveStaticFields and enqueueEverything.

The fix is to split the flag so we can enqueue statics and later enqueue everything.

R=floitsch@google.com, johnniwinther@google.com

Review URL: https://codereview.chromium.org//206013004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@34374 260f80e4-7a28-3924-810f-c04153c831b5