Code transformers fixes for latest analyzer update
This fixes some incompatibilities with the newest analyzer around how deleted files are handled. The resolver was removing the source completely, but the analyzer expects the source to remain as long as it is being referenced by a URI.

This also fixes an issue where resolver were not being released when an exception was thrown from within the applyResolvers call.

** NOTE ** see the commented line which needs to be removed with the updated analyzer.

BUG=
R=brianwilkerson@google.com, scheglov@google.com, sigmund@google.com

Review URL: https://codereview.chromium.org//217163005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@34512 260f80e4-7a28-3924-810f-c04153c831b5